H26-B: Split Compare Table code into multiple components
